+++ lib/src/main/java/com/smileidentity/SmileIDCrashReporting.kt
@@ -27,6 +27,7 @@ object SmileIDCrashReporting {
     fun enable(isInDebugMode: Boolean = false) {
         val options = SentryOptions().apply {
             dsn = BuildConfig.SENTRY_DSN
+            release = BuildConfig.VERSION_NAME
             isEnableUncaughtExceptionHandler = true
             beforeSend = BeforeSendCallback { event: SentryEvent, _: Hint? ->
                 try {