+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Breaking: renamed SmartSelfie Registration to SmartSelfie Enrollment
 - Compile against API level 34
 - Minor change in default colors to add contrast
+- Debounce Selfie Capture directive changes to allow user time to read the directions
+- Tweak selfie progress indicator animation
 
 ### Bug Fixes