+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Minor change in default colors to add contrast
 - Debounce Selfie Capture directive changes to allow user time to read the directions
 - Tweak selfie progress indicator animation
+- Set Release property on Sentry for release tracking
 
 ### Bug Fixes
 - Fix crash on network retries